requirements.txt
numpy
scikit-learn
matplotlib
Pillow
